New Goliath-Envy GUI skin.
Allow application usage without coolbits:
	
	-Any attributes and/ord controllers not readable and controllable from GoliathENVIOUS API will have it's GUI control buttons disabled. Some attributes are still readable.
	
	-If a controller is dependant on another controller and is not the correct value to use the controller, a notification will be shown if the dependant controller is not supported. For example, FanMode.manual is required for setting the target fan speed.
	
	-Specific coolbits values may be used and the GUI will adapt accordingly if the attribute is supported by the GPU. GoliathENVIOUS API groups the following controllers OpertationalStatus(s) together:

		-Core and Memory offset
		-Fan Mode and Fan Target Speed
		-Voltage
	
	-If GoliathENVIOUS API is able to get min-max values for a controller they will show up in GUI regardless of the controllers OperationalStatus. The buttons for setting the value will be disabled, however.
	

Added notifications to provide more user feedback when applying an overclock or fan speed. If GoliathENVIOUS API fails to set the new value the API error message will be shown in the notification.

Added ability to change the location of the OSD, up to 1920x1080.

Added ability to set a new coolbits value(or reset) in GUI if running as root. It is not possible to see the current coolbits value.

Added new Fan Current Speed attribute as using Fan Target Speed is not correct. This change has also been applied to the OSD.

Modified UI to be more consistent. 

Added new warning messages if critical overclocking controllers have an OperationalStatus of NOT_SUPPORTED or READABLE at launch.

-other misc changes.